Fixing codeclimate issues in client.rb
