+++ lib/xclarity_client.rb
@@ -9,11 +9,10 @@ require 'xclarity_client/errors/errors'
 
 require 'xclarity_client/configuration'
 require 'xclarity_client/xclarity_base'
-require 'xclarity_client/xclarity_power_management_mixin'
 require 'xclarity_client/xclarity_credentials_validator'
 require 'xclarity_client/discover'
-require 'xclarity_client/client'
-require 'xclarity_client/schemas'
-
 require 'xclarity_client/endpoints/endpoints'
 require 'xclarity_client/services/services'
+require 'xclarity_client/mixins/mixins'
+require 'xclarity_client/client'
+require 'xclarity_client/schemas'
