+++ lib/xclarity_client/mixins/aicc_mixin.rb
@@ -1,0 +1,10 @@
+module XClarityClient
+  #
+  # Exposes AiccManagement features
+  #
+  module Mixins::AiccMixin
+    def discover_aicc(opts = {})
+      AiccManagement.new(@config).fetch_all(opts)
+    end
+  end
+end
