+++ lib/xclarity_client/mixins/config_target_mixin.rb
@@ -1,0 +1,16 @@
+module XClarityClient
+  #
+  # Exposes ConfigTargetManagement features
+  #
+  module Mixins::ConfigTargetMixin
+    def fetch_config_target(ids = nil,
+                            include_attributes = nil,
+                            exclude_attributes = nil)
+      ConfigTargetManagement.new(@config).get_object_with_id(
+        ids,
+        include_attributes,
+        exclude_attributes
+      )
+    end
+  end
+end
