+++ lib/xclarity_client/mixins/discovery_mixin.rb
@@ -1,0 +1,10 @@
+module XClarityClient
+  #
+  # Exposes DiscoveryManagement features
+  #
+  module Mixins::DiscoverMixin
+    def discover_devices_by_slp
+      DiscoveryManagement.new(@config).fetch_all
+    end
+  end
+end
