+++ lib/xclarity_client/mixins/global_setting_mixin.rb
@@ -1,0 +1,14 @@
+module XClarityClient
+  #
+  # Exposes GlobalSettingManagement features
+  #
+  module Mixins::GlobalSettingMixin
+    def get_globalsettings
+      GlobalSettingManagement.new(@config).population
+    end
+
+    def set_globalsettings(opts = {})
+      GlobalSettingManagement.new(@config).set_globalsettings(opts)
+    end
+  end
+end
