+++ lib/xclarity_client/mixins/host_platform_mixin.rb
@@ -1,0 +1,18 @@
+module XClarityClient
+  #
+  # Exposes HostPlatformManagement features
+  #
+  module Mixins::HostPlatformMixin
+    def get_hostplatforms
+      HostPlatformManagement.new(@config).population
+    end
+
+    def get_osimage_deployment_status(uuid = '')
+      HostPlatformManagement.new(@config).get_osimage_deployment_status(uuid)
+    end
+
+    def deploy_osimage(opts = [])
+      HostPlatformManagement.new(@config).deploy_osimage(opts)
+    end
+  end
+end
