+++ lib/xclarity_client/mixins/mixins.rb
@@ -1,0 +1,39 @@
+module XClarityClient
+  #
+  # Contains mixins to provide XClarityClient::Client features
+  #
+  module Mixins
+  end
+end
+
+require 'xclarity_client/mixins/power_management_mixin'
+require 'xclarity_client/mixins/aicc_mixin'
+require 'xclarity_client/mixins/cabinet_mixin'
+require 'xclarity_client/mixins/canister_mixin'
+require 'xclarity_client/mixins/chassi_mixin'
+require 'xclarity_client/mixins/cmm_mixin'
+require 'xclarity_client/mixins/compliance_policy_mixin'
+require 'xclarity_client/mixins/config_pattern_mixin'
+require 'xclarity_client/mixins/config_profile_mixin'
+require 'xclarity_client/mixins/config_target_mixin'
+require 'xclarity_client/mixins/discover_request_mixin'
+require 'xclarity_client/mixins/discovery_mixin'
+require 'xclarity_client/mixins/event_mixin'
+require 'xclarity_client/mixins/fan_mixin'
+require 'xclarity_client/mixins/fan_mux_mixin'
+require 'xclarity_client/mixins/ffdc_mixin'
+require 'xclarity_client/mixins/global_setting_mixin'
+require 'xclarity_client/mixins/host_platform_mixin'
+require 'xclarity_client/mixins/job_mixin'
+require 'xclarity_client/mixins/node_mixin'
+require 'xclarity_client/mixins/os_image_mixin'
+require 'xclarity_client/mixins/persisted_result_mixin'
+require 'xclarity_client/mixins/power_supply_mixin'
+require 'xclarity_client/mixins/remote_access_mixin'
+require 'xclarity_client/mixins/remote_file_server_mixin'
+require 'xclarity_client/mixins/scalable_complex_mixin'
+require 'xclarity_client/mixins/storage_mixin'
+require 'xclarity_client/mixins/switch_mixin'
+require 'xclarity_client/mixins/unmanage_request_mixin'
+require 'xclarity_client/mixins/update_repo_mixin'
+require 'xclarity_client/mixins/user_mixin'
