+++ lib/xclarity_client/mixins/persisted_result_mixin.rb
@@ -1,0 +1,10 @@
+module XClarityClient
+  #
+  # Exposes PersistedResultManagement features
+  #
+  module Mixins::PersistedResultMixin
+    def fetch_compliance_policies
+      PersistedResultManagement.new(@config).fetch_all
+    end
+  end
+end
