+++ lib/xclarity_client/mixins/power_supply_mixin.rb
@@ -1,0 +1,20 @@
+module XClarityClient
+  #
+  # Exposes PowerSupplyManagement features
+  #
+  module Mixins::PowerSupplyMixin
+    def discover_power_supplies(opts = {})
+      PowerSupplyManagement.new(@config).fetch_all(opts)
+    end
+
+    def fetch_power_supplies(uuids = nil,
+                             include_attributes = nil,
+                             exclude_attributes = nil)
+      PowerSupplyManagement.new(@config).get_object(
+        uuids,
+        include_attributes,
+        exclude_attributes
+      )
+    end
+  end
+end
