+++ lib/xclarity_client/mixins/remote_access_mixin.rb
@@ -1,0 +1,10 @@
+module XClarityClient
+  #
+  # Exposes RemoteAccessManagement features
+  #
+  module Mixins::RemoteAccessMixin
+    def remote_control(uuid)
+      RemoteAccessManagement.new(@config).remote_control(uuid)
+    end
+  end
+end
